storage: change QueueLeaves interface
Shift QueueLeaves to take a slice of QueuedLogLeaf objects so
the caller has to pass in a signature over the LogLeaf as a
commitment that the leaf has been seen.  Storage layer should
store this signature so that it is able to fill in a SET
on retrieval of leaves.

Also move responsibility for filling in the Leaf.QueueTimestamp
to the caller of storage rather than having every storage
implementation have to do this.